Factored out FileListView in separate Class
* Factored out component in sep. class
  (rather first rough cut, optimization to do)
* Fixed missing localization conversion in NavigateMapGo
