+++ FrmMainAppStartup.cs
@@ -196,70 +196,6 @@ public partial class FrmMainApp
     }
 
     /// <summary>
-    ///     applies the names of columnheaders
-    /// </summary>
-    private void AppStartupApplyObjectNames()
-    {
-        Logger.Debug(message: "Starting");
-
-        try
-        {
-            DtObjectNames.DefaultView.Sort = "sqlOrder";
-            DataTable dt = DtObjectNames.DefaultView.ToTable();
-            foreach (DataRow dr in dt.Rows)
-            {
-                ColumnHeader clh = new();
-                string objectName = dr[columnName: "objectName"]
-                    .ToString();
-                clh.Name = "clh_" + objectName;
-                if (lvw_FileList != null)
-                {
-                    lvw_FileList.Columns.Add(value: clh);
-                    Logger.Trace(message: "Loading: " + objectName);
-                }
-                else
-                {
-                    Logger.Error(message: "lvw_FileList is null");
-                }
-            }
-
-            if (lvw_FileList != null)
-            {
-                foreach (ColumnHeader clh in lvw_FileList.Columns)
-                {
-                    clh.Text = HelperStatic.DataReadSQLiteObjectText(
-                        languageName: AppLanguage,
-                        objectType: "ColumnHeader",
-                        objectName: clh.Name
-                    );
-                    Logger.Trace(message: "Loading: " + clh.Name + " --> " + clh.Text);
-                }
-            }
-            else
-            {
-                Logger.Error(message: "lvw_FileList is null");
-            }
-        }
-        catch (Exception ex)
-        {
-            Logger.Fatal(message: "Error: " + ex.Message);
-            MessageBox.Show(text: HelperStatic.GenericGetMessageBoxText(messageBoxName: "mbx_FrmMainApp_ErrorLanguageFileColumnHeaders") + ex.Message, caption: "Error", buttons: MessageBoxButtons.OK, icon: MessageBoxIcon.Error);
-        }
-    }
-
-    /// <summary>
-    ///     Enables column sorting for the main listview
-    /// </summary>
-    private void AppStartupApplyColumnSorting()
-    {
-        // Create the sorter for the list view
-        Logger.Debug(message: "Starting");
-
-        LvwColumnSorter = new ListViewColumnSorter();
-        lvw_FileList.ListViewItemSorter = LvwColumnSorter;
-    }
-
-    /// <summary>
     ///     Sets the startup folder. Defaults to "MyPictures" if null.
     /// </summary>
     private void AppSetupInitialiseStartupFolder()
