+++ Resources/map.html
@@ -66,7 +66,7 @@
 
         { HTMLAddMarker }
 
-        var popup = L.popup()
+        var popup = L.popup();
         var searchControl = L.esri.Geocoding.geosearch({
             position: 'topright',
             placeholder: 'Enter an address or place e.g. 1 York St',
